+++ resources/script.rb
@@ -1,11 +1,11 @@
-actions :run, :delete, :cron
-default_action :cron
+actions :create, :delete
+default_action :create
 
 attribute :name, :kind_of => String, :name_attribute => true
-# cron attributes to map
-attribute :cron, :kind_of => [TrueClass, FalseClass], :default => false
-include Chef::Resource::CronD
-
+attribute :shell, :kind_of => String, :default => 'bash'
+attribute :source, :kind_of => String, :default => 'ec2-consistent-snapshot.erb'
+attribute :cookbook, :kind_of => String, :default => 'ec2-consistent-snapshot'
+attribute :mode, :kind_of => String, :default => '0700'
 # options
 attribute :freeze_filesystem, :kind_of => Array, :default => []
 attribute :region, :kind_of => String, :default => 'us-east-1', :required => true
@@ -28,7 +28,7 @@ attribute :mongo_password, :kind_of => [String, NilClass], :default => nil
 # mysql
 attribute :mysql, :kind_of => [TrueClass, FalseClass], :default => true
 attribute :mysql_defaults_file, :kind_of => [String, NilClass],  :default => nil
-attribute :mysql_host, :kind_of => [String, NilClass], :default => 'localhost'
+attribute :mysql_host, :kind_of => [String, NilClass], :default => nil
 attribute :mysql_port, :kind_of => [String, Integer, NilClass], :default => 3306
 
 attribute :mysql_socket, :kind_of => [String, NilClass], :default => nil